Collapse "upload-image" into "setup-page" again
Now that it's working I'm moving back toward putting it where it'll be used. I'll keep it as a separate component because I know for sure we'll use it in a few places
